ifex_ast_doc: Handle ForwardRef by skipping it
It was noticed that if we start with AST as the top node when walking
through all Node types, then AST has a ForwardRef which caused a
runtime error.

The purpose of the ForwardRef is to reference a not yet defined class.
It should be safe to skip it because in all those cases we would have a
reference to the class somewhere else (otherwise no need for the
ForwardRef trick - we could just rearrange the order of definitions in
that case)

For the documentation generation there is no difference because there we
start with Namespace as the top node, but the same functions will be
reused in the code that matches Node class types to corresponding
template file names.  It seems more correct to start with the topmost
node (AST) when traversing the node type tree.