Introduce TemplateDir module
A module to collect up functions and data dealing with the template
definitions.

Adds initial support for template subdirectories below /templates
(what was before the module-global variable TemplatePath)

Adds new functions to find templates in a subdirectory and match
them by-name against the node class types.  The AST node classes are
found by traversing the ifex_ast dataclasses, by reusing the code that
was written for the documentation generation.  Instead of a hard-coded
default_templates variable, it is now more possible to specify a
directory and find matching templates by name, in it.

Except for the new functions, this moves already existing code that was
in the anonymous __init__.py file into the name TemplateDir. That aspect
does little except wrap the module in another name.

It is however a step towards, maybe, making this an instantiable class
instead, which could support handling more than one template
subdirectory and a way to specify the wanted subdirectory at run-time
instead of load-time.

For now however, it is still a single-instance module setup without any
instantiable class.

Signed-off-by: Gunnar Andersson <gunnar_dev@nospam@novaspring.eu>